dag-scrape-project/requirements.txt
agate==1.6.3
alembic==1.10.2
aniso8601==9.0.1
anyio==3.6.2
attrs==22.2.0
Babel==2.12.1
backoff==2.2.1
beautifulsoup4==4.12.0
betterproto==1.2.5
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coloredlogs==14.0
commonmark==0.9.1
croniter==1.3.8
dagit==1.2.4
dagster==1.2.4
dagster-dbt==0.18.4
dagster-graphql==1.2.4
dbt-bigquery==1.4.3
dbt-core==1.4.5
dbt-extractor==0.4.1
docstring-parser==0.15
fsspec==2023.3.0
future==0.18.3
google-api-core==2.11.0
google-auth==2.17.1
google-cloud-bigquery==3.9.0
google-cloud-core==2.3.2
google-cloud-dataproc==5.4.1
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
gql==3.4.0
graphene==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.47.5
grpcio-health-checking==1.47.5
grpcio-status==1.51.3
grpclib==0.4.3
h11==0.14.0
h2==4.1.0
hologram==0.0.15
hpack==4.0.0
httptools==0.5.0
humanfriendly==10.0
hyperframe==6.0.1
idna==3.4
isodate==0.6.1
Jinja2==3.1.2
jsonschema==3.2.0
leather==0.3.4
Logbook==1.5.3
Mako==1.2.4
MarkupSafe==2.1.2
mashumaro==3.3.1
minimal-snowplow-tracker==0.0.2
msgpack==1.0.5
multidict==6.0.4
networkx==2.8.8
packaging==23.0
parsedatetime==2.4
pathspec==0.10.3
pendulum==2.1.2
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.7
Pygments==2.14.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytimeparse==1.1.8
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
requests==2.28.2
requests-toolbelt==0.10.1
rich==12.6.0
rsa==4.9
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==1.4.47
sqlparse==0.4.3
starlette==0.26.1
stringcase==1.2.0
tabulate==0.9.0
text-unidecode==1.3
tomli==2.0.1
toposort==1.10
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
universal_pathlib==0.0.23
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchdog==3.0.0
watchfiles==0.19.0
websockets==10.4
Werkzeug==2.2.3
yarl==1.8.2
